refactoring statistic code removing duplicates
